renovate code club screen

+++ project_api/src/main/java/com/example/project_api/models/tables/Time.java
@@ -15,8 +15,8 @@ public class Time {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
-    private int fieldId;
-    private int userId;
+    private long fieldId;
+    private long userId;
     private String startTime;
     private String endTime;
     private boolean status;
